+++ wallet-sdk/src/androidMain/kotlin/com/algorand/common/algosdk/AlgoSdkUtils.android.kt
@@ -10,24 +10,9 @@
  * limitations under the License
  */
 
-package com.algorand.common.encryption
+package com.algorand.common.algosdk
 
-import kotlin.io.encoding.Base64
-import kotlin.io.encoding.ExperimentalEncodingApi
-
-internal actual fun getBase64Manager(): Base64Manager {
-    return Base64ManagerImpl()
-}
-
-internal class Base64ManagerImpl : Base64Manager {
-
-    @OptIn(ExperimentalEncodingApi::class)
-    override fun encode(byteArray: ByteArray): String {
-        return Base64.encode(byteArray)
-    }
-
-    @OptIn(ExperimentalEncodingApi::class)
-    override fun decode(value: String): ByteArray {
-        return Base64.decode(value)
-    }
+actual fun isValidAccountAddress(address: String): Boolean {
+    // TODO Use algo sdk
+    return Regex("^([A-Z2-7]{58})$").matches(address)
 }
