+++ wallet-sdk/src/commonMain/kotlin/com/algorand/common/encryption/Base64Manager.kt
@@ -18,5 +18,3 @@ interface Base64Manager {
 
     fun decode(value: String): ByteArray
 }
-
-internal expect fun getBase64Manager(): Base64Manager
